Allow executables to be mapped to a virtual memory address. Now (finally) multiple executables can share the same logical address, and the GUI shell supports displaying the output of multiple programs.

+++ MyOS_1/main.c
@@ -118,7 +118,7 @@ void KeStartupPhase2(multiboot_info *multibootInfo)
     tasks[0].inUse = true;
     strncpy(tasks[0].imageName, "KERNEL PROCESS", sizeof("KERNEL PROCESS"));
     tasks[0].PID = nextPID - 1;
-    tasks[0].cr3 = pageDir;
+    tasks[0].cr3 = (uint32_t)pageDir;
 
     // Initialize interrupts
     Interrupts_Init();
